+++ mypy.ini
@@ -1,2 +1,2 @@
 [mypy]
-exclude = bandcrash/gui.py
+exclude = bandcrash/gui
